Add new options to pipeline

+++ source/CASPRg.sh
@@ -23,6 +23,8 @@ Usage: generate_counts.sh [-h or --help]
                           [-i or --info-alignment]
                           [-k or --keep-tmp]
                           [-q or --output-folder]
+                          [-s or --start]
+                          [-p or --pause]
 """
 }
 function print_error {
@@ -69,7 +71,7 @@ Optional arguments:
     -b, --bases-aligned:
                 Minimum number of bases that must be aligned for counts.
                 It must be an integer.
-                Default: 20 for sgRNAs and 35 for pgRNAs.
+                Default: 20 for sgRNAs and 30 for pgRNAs.
     -y, --fdr-threshold:
                 FDR threshold for the identification of hits.
                 Must be a number between 0 and 1 (limits not included).
@@ -91,6 +93,15 @@ Optional arguments:
                 Path and name of the directory to save outputs.
                 The directory must exist. It will not be created.
                 Default: current directory.
+    -s, --start:
+                Step from which to start the analysis.
+                Options: qc, trim, map, test.
+                Default: start indexing the genome (1st step).
+    -p, --pause:
+                Step at which to stop the analysis.
+                This step will be included in the execution.
+                Options: indexing, qc, trim, map.
+                Default: stop when the test is finished (last step).
 
 Required arguments:
     -f, --fastq-forward:
@@ -128,6 +139,11 @@ Important:
     The number of input files in -f and -r must be the same (for pgRNAs).
     The SampleName's of -f and -r files must be identical.
     Forward and reverse fastq files must be provided in the same order.
+
+    If --start test, only --exper-design is mandatory, with --output-dir,
+    --fdr-threshold and --controls --output-dir being optional.
+    In this case, file named output-dir/outputs/table.counts.txt MUST exist.
+    The names of its columns must correspond to SampleNames in --exper-design.
 """
 }
 
@@ -150,6 +166,8 @@ for ARGS in "$@"; do
                 "--info-alignment") set -- "$@" "-i" ;;
                 "--keep-tmp") set -- "$@" "-k" ;;
                 "--output-dir") set -- "$@" "-q" ;;
+                "--start") set -- "$@" "-s" ;;
+                "--pause") set -- "$@" "-p" ;;
                 "--help") set -- "$@" "-h" ;;
                 *) set - "$@" "$ARGS"
         esac
@@ -159,7 +177,7 @@ done
 o=53; a="ACCG"; A="AAAC"; m=0; b=0; t=1; i=0; k=0; y=0.1; q="./"
 
 # Define all parameters
-while getopts 'r:f:l:e:c:o::a::A::m::b::y::t::q::ikh' flag; do
+while getopts 'r:f:l:e:c:o::a::A::m::b::y::t::q::s:p:ikh' flag; do
         case "${flag}" in
                 r) r=${OPTARG} ;;
                 f) f=${OPTARG} ;;
@@ -174,6 +192,8 @@ while getopts 'r:f:l:e:c:o::a::A::m::b::y::t::q::ikh' flag; do
                 y) y=${OPTARG} ;;
                 t) t=${OPTARG} ;;
                 q) q=${OPTARG} ;;
+                s) start=${OPTARG} ;;
+                p) pause=${OPTARG} ;;
                 i) i=1 ;;
                 k) k=1 ;;
                 h) print_help
@@ -193,7 +213,7 @@ if [[ $r == "" ]]; then
   i=0
   if [[ $b == 0 ]]; then b=20; fi
 else
-  if [[ $b == 0 ]]; then b=35; fi
+  if [[ $b == 0 ]]; then b=30; fi
 fi
 
 ########################################
@@ -211,10 +231,16 @@ if [[ $(echo "${q: -1}") != "/" ]]; then
   q="$q/"
 fi
 
-# Create all folders that will be needed
-rm -rf "${q}/intermediate" && \
-mkdir "${q}/intermediate" && chmod +xwr "${q}/intermediate"
-rm -rf "${q}/outputs" && mkdir "${q}/outputs" && chmod +xwr "${q}/outputs"
+# Create all folders that will be needed inside output directory
+if [[ $start == "" || $start == "qc" ]]; then
+  rm -rf "${q}/intermediate" && \
+  mkdir "${q}/intermediate" && chmod +xwr "${q}/intermediate"
+  rm -rf "${q}/outputs" && \
+  mkdir "${q}/outputs" && chmod +xwr "${q}/outputs"
+else
+  mkdir -p "${q}/intermediate" && chmod +xwr "${q}/intermediate"
+  mkdir -p "${q}/outputs" && chmod +xwr "${q}/outputs"
+fi
 
 
 ########################################
@@ -225,71 +251,98 @@ rm -rf "${q}/outputs" && mkdir "${q}/outputs" && chmod +xwr "${q}/outputs"
 currentdir=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )
 
 # Display any error that is initially detected
-bash $currentdir/basic_errors.sh \
-  "$f" "$r" $l $e $o $a $A $m $b $y $t $q "$c" # Inputs
-if [[ $(echo $?) != 0 ]]; then exit 1; fi # Exit if there has been an error.
+if [[ $start != "test" ]]; then
+  bash $currentdir/basic_errors.sh \
+    "$f" "$r" $l $e $o $a $A $m $b $y $t $q "$c" "$start" "$pause" # Inputs
+  if [[ $(echo $?) != 0 ]]; then exit 1; fi # Exit if there has been an error.
+else
+  bash $currentdir/basic_errors_test.sh \
+    $e $y $q "$c" # Inputs
+  if [[ $(echo $?) != 0 ]]; then exit 1; fi # Exit if there has been an error.
+fi
 
 # Create a fake genome with the guide RNAs
-bash $currentdir/fake_genome.sh $l $t $q 2> "${q}/intermediate/err.txt"
-if [[ $(cat "${q}/intermediate/err.txt") != "" ]]; then # Exit if error
-  echo "Error: problems with the library file. Check it again."
-  echo "See ${q}/intermediate/err.txt for more information"
+if [[ $start == "" ]]; then
+  bash $currentdir/fake_genome.sh $l $t $q 2> "${q}/intermediate/err.txt"
+  if [[ $(cat "${q}/intermediate/err.txt") != "" ]]; then # Exit if error
+    echo "Error: problems with the library file. Check it again."
+    echo "See ${q}/intermediate/err.txt for more information"
+    exit 1
+  fi
+fi
+if [[ $pause == "indexing" ]]; then
   exit 1
 fi
 
 # Perform QC of the fastq files. No new errors are expected in this step
-bash $currentdir/quality_control.sh \
-      "$f" "$r" $t $q 2> "${q}/intermediate/err.txt"
+if [[ $start == "" || $start == "qc" ]]; then
+  bash $currentdir/quality_control.sh \
+        "$f" "$r" $t $q 2> "${q}/intermediate/err.txt"
+fi
+if [[ $pause == "qc" ]]; then
+  exit 1
+fi
 
 # Trimming and arrangement of the reads
-bash $currentdir/trimming_reads.sh \
-          "$f" "$r" $l $o $a $A $q $t 2> "${q}/intermediate/err.txt"
-if [[ $(cat "${q}/intermediate/err.txt") != "" ]]; then # Exit if error
-  echo "Error: problems with the input parameters. Check them again."
-  echo "See ${q}/intermediate/err.txt for more information"
+if [[ $start == "" || $start == "qc" || $start == "trim" ]]; then
+  bash $currentdir/trimming_reads.sh \
+            "$f" "$r" $l $o $a $A $q $t 2> "${q}/intermediate/err.txt"
+  if [[ $(cat "${q}/intermediate/err.txt") != "" ]]; then # Exit if error
+    echo "Error: problems with the input parameters. Check them again."
+    echo "See ${q}/intermediate/err.txt for more information"
+    exit 1
+  fi
+
+  # Create plots with the trimming statistics
+  printf "\nGenerating plots to see trimming statistics\n"
+  # Collect needed information previously created
+  triminf=$(cat "${q}/intermediate/useful_information.txt" | \
+            awk 'NR>1' | sort -k1 | awk '{print $2}')
+  # Run R script
+  Rscript --vanilla $currentdir/trimming_statistics.R \
+            "$triminf" "$q" ${q}/intermediate/trim_stat* \
+  || (echo "Problem with R. Check if the R version is correct." && exit 1)
+  if [[ $(echo $?) != 0 ]]; then exit 1; fi # Exit if there has been an error.
+  echo "Plots were created successfully"
+fi
+if [[ $pause == "trim" ]]; then
   exit 1
 fi
-
-# Create plots with the trimming statistics
-printf "\nGenerating plots to see trimming statistics\n"
-# Collect needed information previously created
-triminf=$(cat "${q}/intermediate/useful_information.txt" | \
-          awk 'NR>1' | sort -k1 | awk '{print $2}')
-# Run R script
-Rscript --vanilla $currentdir/trimming_statistics.R \
-          "$triminf" "$q" ${q}/intermediate/trim_stat* \
-|| (echo "Problem with R. Check if the R version is correct." && exit 1)
-if [[ $(echo $?) != 0 ]]; then exit 1; fi # Exit if there has been an error.
-echo "Plots were created successfully"
 
 # Align the reads to the fake genome and count the number of reads per guide
-bash $currentdir/alignment_counts.sh \
-      $m $b $t $q "$r" $l $i 2> "${q}/intermediate/err.txt"
-if [[ $(cat "${q}/intermediate/err.txt") != "" ]]; then # Exit if error
-  echo "Error: problems with the input parameters. Check them again."
-  echo "See ${q}/intermediate/err.txt for more information."
-  exit 1
-fi
-
-# Create plots with the alignment statistics
-printf "\nGenerating plots to see alignment statistics\n"
-Rscript --vanilla $currentdir/alignment_statistics.R \
-        $q ${q}/intermediate/Statistics_alignment* \
-|| (echo "Problem with R. Check if the R version is correct." && exit 1)
-if [[ $(echo $?) != 0 ]]; then exit 1; fi # Exit if there has been an error.
-
-# Create plots to see what happens with unmapped reads (only if requested)
-if [[ $i == 1 ]]; then
-  bash $currentdir/alignment_info.sh $q "$f" $b 2> "${q}/intermediate/err.txt"
+if [[ $start != "test" ]]; then
+  bash $currentdir/alignment_counts.sh \
+        $m $b $t $q "$r" $l $i 2> "${q}/intermediate/err.txt"
   if [[ $(cat "${q}/intermediate/err.txt") != "" ]]; then # Exit if error
     echo "Error: problems with the input parameters. Check them again."
     echo "See ${q}/intermediate/err.txt for more information."
     exit 1
   fi
-else
-  mv ${q}/intermediate/Alignment_statistics.pdf ${q}/outputs
+
+  # Create plots with the alignment statistics
+  printf "\nGenerating plots to see alignment statistics\n"
+  Rscript --vanilla $currentdir/alignment_statistics.R \
+          $q ${q}/intermediate/Statistics_alignment* \
+  || (echo "Problem with R. Check if the R version is correct." && exit 1)
+  if [[ $(echo $?) != 0 ]]; then exit 1; fi # Exit if there has been an error.
+
+  # Create plots to see what happens with unmapped reads (only if requested)
+  if [[ $i == 1 ]]; then
+    bash $currentdir/alignment_info.sh \
+              $q "$f" $b $l $currentdir 2> "${q}/intermediate/err.txt"
+    if [[ $(cat "${q}/intermediate/err.txt") != "" ]]; then # Exit if error
+      echo "Error: problems with the input parameters. Check them again."
+      echo "See ${q}/intermediate/err.txt for more information."
+      exit 1
+    fi
+  else
+    mv ${q}/intermediate/Alignment_statistics.pdf ${q}/outputs
+  fi
+  printf "Plots were created successfully\n"
 fi
-printf "\nPlots were created successfully\n"
+if [[ $pause == "map" ]]; then
+  exit 1
+fi
 
 # Compute the test to see the significant genes.
 bash \
@@ -310,7 +363,7 @@ if [[ $(echo $?) != 0 ]]; then exit 1; fi # Exit if there has been an error.
 echo "Plots were created successfully"
 
 # Create files for the visualization with VISPR
-# bash $currentdir/visualization.sh ${q} $currentdir
+bash $currentdir/visualization.sh ${q} $currentdir
 
 ###############################
 ## Remove intermediate files ##
